+++ src/init.c
@@ -17,6 +17,8 @@ extern SEXP zstd_unserialize_stream_file_(SEXP raw_vec_, SEXP dctx_);
 extern SEXP zstd_serialize_stream_(SEXP robj_, SEXP level_, SEXP num_threads_, SEXP cctx_);
 extern SEXP zstd_unserialize_stream_(SEXP raw_vec_, SEXP dctx_);
 
+extern SEXP zstd_train_dictionary_(SEXP samples_, SEXP size_, SEXP optim_, SEXP optim_shrink_allow_);
+extern SEXP zstd_dict_id_(SEXP dict_);
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // .C      R_CMethodDef
@@ -37,6 +39,9 @@ static const R_CallMethodDef CEntries[] = {
   
   {"zstd_serialize_stream_"       , (DL_FUNC) &zstd_serialize_stream_       , 4},
   {"zstd_unserialize_stream_"     , (DL_FUNC) &zstd_unserialize_stream_     , 2},
+  
+  {"zstd_train_dictionary_"       , (DL_FUNC) &zstd_train_dictionary_       , 4},
+  {"zstd_dict_id_"                , (DL_FUNC) &zstd_dict_id_                , 2},
   
   {NULL, NULL, 0}
 };